+++ src/OneImlx.Terminal/Commands/Checkers/DataValidationValueChecker.cs
@@ -39,7 +39,7 @@ namespace OneImlx.Terminal.Commands.Checkers
         }
 
         /// <inheritdoc/>
-        public override bool Equals(object obj)
+        public override bool Equals(object? obj)
         {
             if (obj == null)
             {