+++ test/OneImlx.Terminal.Tests/Commands/Parsers/CommandRouteParserTests.cs
@@ -37,7 +37,7 @@ namespace OneImlx.Terminal.Commands.Parsers
                new CommandDescriptor("cmd_nr1", "cmd_nr1_name", "cmd_nr1_desc", CommandType.SubCommand, CommandFlags.None),
                new CommandDescriptor("cmd_nr2", "cmd_nr2_name", "cmd_nr2_desc", CommandType.SubCommand, CommandFlags.None)
             });
-            commandStore = new TerminalInMemoryImmutableCommandStore(textHandler, commandDescriptors.Values);
+            commandStore = new TerminalInMemoryCommandStore(textHandler, commandDescriptors.Values);
             logger = new NullLogger<CommandRouteParser>();
 
             commandRouteParser = new CommandRouteParser(textHandler, commandStore, terminalOptions, logger);
@@ -409,7 +409,7 @@ namespace OneImlx.Terminal.Commands.Parsers
 
         private readonly TerminalOptions terminalOptions;
         private ITerminalTextHandler textHandler;
-        private ITerminalImmutableCommandStore commandStore;
+        private ITerminalCommandStore commandStore;
         private CommandDescriptors commandDescriptors;
         private ICommandRouteParser commandRouteParser;
         private ILogger<CommandRouteParser> logger;